+++ src/app/modules/organizations/data-quality/inventory/modal/delete-modal.component.html
